+++ search-films/src/pages/MainPage/index.tsx
@@ -1,34 +1,13 @@
 
-import { useLoaderData, useNavigate } from 'react-router-dom';
+import { useLoaderData } from 'react-router-dom';
 import { FilmDescApi } from '../../api';
-import { Flex, Input, MovieCard, Paragraph, Title } from '../../components';
+import { Flex, MovieCard } from '../../components';
 
 
 function MainPage() {
 	const data = useLoaderData() as FilmDescApi[];
-	const navigate = useNavigate();
-
-	const saveTextToSearch = (textToSearch: string) => {
-		if (textToSearch) {
-			navigate('/' + textToSearch);
-		} 
-	};
-
 	return (
 		<>
-			<Title text='Поиск'/>
-			<Paragraph 
-				text='Введите название фильма, сериала или мультфильма для поиска и добавления в избранное.'
-				fontSizeInPx={16}
-			/>
-			<Input 
-				position='vertical'
-				placeholder="Введите название"
-				onSend={(textToSearch: string) => {saveTextToSearch(textToSearch);}}
-				inputActionName="Искать"
-				iconClassName={'icon-search'}
-				hasIcon={true}
-			/>
 			<Flex position='horizontal' paddingTop={88}>
 				{
 					data && data.map((d: FilmDescApi) => {