+++ search-films/src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, defer, RouterProvider } from 'react-router-dom';
-import { doSearchFilmDescByName } from './api';
+import { doGetDetails, doSearchFilmDescByName } from './api';
 import { UserProvider } from './context/user.context';
 import './index.css';
 import MenuLayout from './layouts/MenuLayout';
@@ -9,6 +9,7 @@ import SearchLayout from './layouts/SearchLayout';
 import ErrorPage from './pages/ErrorPage/index.js';
 import FavoritesPage from './pages/FavoritesPage/index.js';
 import { LoginPage, MainPage } from './pages/index.js';
+import MoviePage from './pages/MoviePage';
 import SearchPage from './pages/SearchPage';
 
 const router = createBrowserRouter([
@@ -29,6 +30,18 @@ const router = createBrowserRouter([
 						throw new Error('Invalid movie name');
 					}
 					return defer({data: doSearchFilmDescByName(params.movieName)});
+
+				}
+			},
+			{
+				path: '/movie/:movieId',
+				element: <MoviePage/>,
+				errorElement: <ErrorPage/>,
+				loader: async ({ params }) => {
+					if (params.movieId == undefined) {
+						throw new Error('Invalid movie id');
+					}
+					return defer({data: doGetDetails(params.movieId)});
 
 				}
 			},