+++ search-films/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ##### New Features
 
-- update routes (be32f075)
+- add Loader's support (b8b1dc81)
+- update routes (09340df1)
 - impl of requests (mock, real) (9cca891b)
 
 ##### Bug Fixes